requirements.txt
langchain
langchain-openai
python-dotenv
openai
pytest
PyPDF2
python-docx
